refine serialization tests
