+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/keras/nn/DenseSpec.scala
@@ -22,6 +22,9 @@ import com.intel.analytics.bigdl.nn.abstractnn.AbstractModule
 import com.intel.analytics.bigdl.nn.keras.Dense
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.utils.Shape
+import com.intel.analytics.bigdl.utils.serializer.ModuleSerializationTest
+
+import scala.util.Random
 
 class DenseSpec extends KerasBaseSpec {
 
@@ -60,4 +63,13 @@ class DenseSpec extends KerasBaseSpec {
       kerasCode, weightConverter, precision = 1e-4)
   }
 
+}
+
+class DenseSerialTest extends ModuleSerializationTest {
+  override def test(): Unit = {
+    val dense = Dense[Float](10, inputShape = Shape(20))
+    dense.build(Shape(2, 20))
+    val input = Tensor[Float](2, 20).apply1(_ => Random.nextFloat())
+    runSerializationTest(dense, input)
+  }
 }
