+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/FlattenTableSpec.scala
@@ -16,8 +16,11 @@
 package com.intel.analytics.bigdl.nn
 
 import com.intel.analytics.bigdl.tensor.{Storage, Tensor}
-import com.intel.analytics.bigdl.utils.T
+import com.intel.analytics.bigdl.utils.{T, Table}
+import com.intel.analytics.bigdl.utils.serializer.ModuleSerializationTest
 import org.scalatest.{BeforeAndAfter, FlatSpec, Matchers}
+
+import scala.util.Random
 
 @com.intel.analytics.bigdl.tags.Parallel
 class FlattenTableSpec extends FlatSpec with BeforeAndAfter with Matchers {
@@ -124,3 +127,14 @@ class FlattenTableSpec extends FlatSpec with BeforeAndAfter with Matchers {
   }
 }
 
+class FlattenTableSerialTest extends ModuleSerializationTest {
+  override def test(): Unit = {
+    val flattenTable = FlattenTable[Float]().setName("flattenTable")
+    val input1 = Tensor[Float](5, 5).apply1(e => Random.nextFloat())
+    val input2 = Tensor[Float](5, 5).apply1(e => Random.nextFloat())
+    var input = new Table()
+    input(1.toFloat) = input1
+    input(2.toFloat) = input2
+    runSerializationTest(flattenTable, input)
+  }
+}
