+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/RecurrentDecoderSpec.scala
@@ -22,10 +22,12 @@ import com.intel.analytics.bigdl.optim.{L2Regularizer, SGD}
 import com.intel.analytics.bigdl.utils._
 import com.intel.analytics.bigdl.tensor.{Storage, Tensor}
 import com.intel.analytics.bigdl.utils.RandomGenerator._
+import com.intel.analytics.bigdl.utils.serializer.ModuleSerializationTest
 import org.scalatest.{BeforeAndAfter, FlatSpec, Matchers}
 
 import scala.collection.mutable.ArrayBuffer
 import scala.math._
+import scala.util.Random
 
 @com.intel.analytics.bigdl.tags.Parallel
 class RecurrentDecoderSpec extends FlatSpec with BeforeAndAfter with Matchers {
@@ -401,5 +403,14 @@ class RecurrentDecoderSpec extends FlatSpec with BeforeAndAfter with Matchers {
       assert(abs(v1 - v2) <= 1e-8)
       v1
     })
+  }
+}
+
+class RecurrentDecoderSerialTest extends ModuleSerializationTest {
+  override def test(): Unit = {
+    val recDecoder = RecurrentDecoder[Float](5).
+      add(ConvLSTMPeephole[Float](7, 7, 3, 3, 1))
+    val input = Tensor[Float](4, 7, 5, 5).apply1(_ => Random.nextFloat())
+    runSerializationTest(recDecoder, input)
   }
 }
