+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/ops/Dilation2DSpec.scala
@@ -18,32 +18,14 @@ package com.intel.analytics.bigdl.nn.ops
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.utils.T
 import com.intel.analytics.bigdl.utils.serializer.ModuleSerializationTest
-import org.scalatest.{FlatSpec, Matchers}
 
-import scala.util.Random
-
-class ProdSpec extends FlatSpec with Matchers {
-  "Prod operation" should "works correctly" in {
-    import com.intel.analytics.bigdl.numeric.NumericFloat
-    val input =
-      Tensor(T(
-        T(1f, 2f, 3f),
-        T(2f, 2f, 4f),
-        T(2f, 2f, 4f)
-      ))
-
-    val expectOutput = Tensor(T(4f, 8f, 48f))
-
-    val output = Prod(axis = 1).forward(input)
-    output should be(expectOutput)
-  }
-}
-
-class ProdSerialTest extends ModuleSerializationTest {
+class Dilation2DSerialTest extends ModuleSerializationTest {
   override def test(): Unit = {
-    val prod = Prod[Float](-1, false).setName("prod")
-    val input = Tensor[Float](3, 3).apply1(_ => Random.nextFloat())
-    runSerializationTest(prod, input, prod.
-      asInstanceOf[ModuleToOperation[Float]].module.getClass)
+    val module = Dilation2D[Float, Float](
+      Array(1, 3, 2, 1), Array(1, 2, 3, 1), "same")
+
+    val input = T(Tensor[Float](4, 32, 32, 3).rand(), Tensor[Float](3, 4, 3).rand())
+
+    runSerializationTest(module, input)
   }
 }
