+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/ops/StackOpsSpec.scala
@@ -15,35 +15,24 @@
  */
 package com.intel.analytics.bigdl.nn.ops
 
+import com.intel.analytics.bigdl.nn.Graph
+import com.intel.analytics.bigdl.nn.tf.Const
 import com.intel.analytics.bigdl.tensor.Tensor
-import com.intel.analytics.bigdl.utils.T
 import com.intel.analytics.bigdl.utils.serializer.ModuleSerializationTest
-import org.scalatest.{FlatSpec, Matchers}
 
-import scala.util.Random
-
-class ProdSpec extends FlatSpec with Matchers {
-  "Prod operation" should "works correctly" in {
-    import com.intel.analytics.bigdl.numeric.NumericFloat
-    val input =
-      Tensor(T(
-        T(1f, 2f, 3f),
-        T(2f, 2f, 4f),
-        T(2f, 2f, 4f)
-      ))
-
-    val expectOutput = Tensor(T(4f, 8f, 48f))
-
-    val output = Prod(axis = 1).forward(input)
-    output should be(expectOutput)
-  }
-}
-
-class ProdSerialTest extends ModuleSerializationTest {
+class StackOpsSerialTest extends ModuleSerializationTest {
   override def test(): Unit = {
-    val prod = Prod[Float](-1, false).setName("prod")
-    val input = Tensor[Float](3, 3).apply1(_ => Random.nextFloat())
-    runSerializationTest(prod, input, prod.
-      asInstanceOf[ModuleToOperation[Float]].module.getClass)
+    val data = Const[Float, Float](Tensor.scalar[Float](1)).inputs()
+    val stack = new StackCreator[Float, Float]().inputs()
+    val push = new StackPush[Float, Float]().inputs(stack, data)
+    val ctr = new com.intel.analytics.bigdl.nn.tf.ControlDependency[Float]().inputs(push)
+    val pop = new StackPop[Float, Float]().inputs(stack, ctr)
+    val model = Graph.dynamic[Float](Array(stack), Array(pop))
+
+    runSerializationTestWithMultiClass(model, Tensor.scalar(1), Array(
+      stack.element.getClass.asInstanceOf[Class[_]],
+      push.element.getClass.asInstanceOf[Class[_]],
+      pop.element.getClass.asInstanceOf[Class[_]]
+    ))
   }
 }
