+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/utils/tf/loaders/SplitLoadTFSpec.scala
@@ -13,37 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn.ops
+package com.intel.analytics.bigdl.utils.tf.loaders
 
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.utils.T
 import com.intel.analytics.bigdl.utils.serializer.ModuleSerializationTest
-import org.scalatest.{FlatSpec, Matchers}
 
 import scala.util.Random
 
-class ProdSpec extends FlatSpec with Matchers {
-  "Prod operation" should "works correctly" in {
-    import com.intel.analytics.bigdl.numeric.NumericFloat
-    val input =
-      Tensor(T(
-        T(1f, 2f, 3f),
-        T(2f, 2f, 4f),
-        T(2f, 2f, 4f)
-      ))
-
-    val expectOutput = Tensor(T(4f, 8f, 48f))
-
-    val output = Prod(axis = 1).forward(input)
-    output should be(expectOutput)
-  }
-}
-
-class ProdSerialTest extends ModuleSerializationTest {
+class SplitLoadTFSerialTest extends ModuleSerializationTest {
   override def test(): Unit = {
-    val prod = Prod[Float](-1, false).setName("prod")
-    val input = Tensor[Float](3, 3).apply1(_ => Random.nextFloat())
-    runSerializationTest(prod, input, prod.
-      asInstanceOf[ModuleToOperation[Float]].module.getClass)
+    val splitLoadTF = new SplitLoadTF[Float](1).setName("splitLoadTD")
+    val input = T(Tensor[Int](T(1)),
+      Tensor[Float](1, 6, 2).apply1(_ => Random.nextFloat())
+    )
+    runSerializationTest(splitLoadTF, input)
   }
 }
